+++ Assets/OxGKit/InputSystem/Scripts/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [0.0.2-preview] - 2023-08-27
 - Added Logger by LoggingSystem.
+- Renamed Inputs.IA.OnUpdateInputActions to Inputs.IA.UpdateInputActions.
 
 ## [0.0.1-preview] - 2023-05-21
 - Added InputSystem.